apps/control-plane/requirements.txt
boto3==1.28.62
botocore==1.31.62
psycopg2-binary>=2.9.5
requests==2.31.0
